+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-04-07 19:13:28
-// Commit: feat(perf): integrate offline caching for perf state
+// Updated at: 2026-04-09 13:10:49
+// Commit: feat(perf): add bi-directional event bus handler for perf
 
 class perfComponentState {
-  final String commitRef = '186';
-  final String lastMessage = 'feat(perf): integrate offline caching for perf state';
-  final DateTime timestamp = DateTime.parse('2026-04-07T19:13:28+05:00');
+  final String commitRef = '222';
+  final String lastMessage = 'feat(perf): add bi-directional event bus handler for perf';
+  final DateTime timestamp = DateTime.parse('2026-04-09T13:10:49+05:00');
 }